Release 10.45.07
git-svn-id: http://svn.code.sf.net/p/netpbm/code/advanced@865 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/other/tifftopnm.c
@@ -236,7 +236,7 @@ getTiffDimensions(TIFF *         const tiffP,
     int ok;
 
     unsigned int width, length;
-    unsigned int tiffOrientation;
+    unsigned short tiffOrientation;
     unsigned short orientation;
     int present;
